+++ Misc/Misc.h
@@ -44,7 +44,7 @@ namespace CMISC {
     inline double Tic() {
         struct timeval tv;
         gettimeofday(&tv, NULL);
-        return  tv.tv_sec + 1e-6 * (tv.tv_usec);
+        return static_cast<double>( tv.tv_sec ) + 1e-6 * static_cast<double>(tv.tv_usec);
     }
 
     ////////////////////////////////////////////////////////////////////////////////
